+++ src/graphql_clj/validator.clj
@@ -14,6 +14,7 @@
             [graphql-clj.validator.rules.unique-input-field-names]
             [graphql-clj.validator.rules.unique-fragment-names]
             [graphql-clj.validator.rules.unique-argument-names]
+            [graphql-clj.validator.rules.provided-non-null-arguments]
             [graphql-clj.visitor :as visitor]
             [graphql-clj.spec :as spec]
             [instaparse.core :as insta]
@@ -21,12 +22,16 @@
 
 (def first-pass-rules [spec/add-spec spec/define-specs])
 
-(def second-pass-rules
+(def second-pass-rules-schema
+  (flatten [graphql-clj.validator.rules.unique-input-field-names/schema-rules
+            graphql-clj.validator.rules.unique-argument-names/schema-rules]))
+
+(def second-pass-rules-statement
   (flatten [graphql-clj.validator.rules.known-type-names/rules
             graphql-clj.validator.rules.known-argument-names/rules
             graphql-clj.validator.rules.known-fragment-names/rules
             graphql-clj.validator.rules.no-undefined-variables/rules
-            graphql-clj.validator.rules.unique-input-field-names/rules
+            graphql-clj.validator.rules.unique-input-field-names/statement-rules
             graphql-clj.validator.rules.arguments-of-correct-type/rules
             graphql-clj.validator.rules.default-values-of-correct-type/rules
             graphql-clj.validator.rules.variables-are-input-types/rules
@@ -36,7 +41,8 @@
             graphql-clj.validator.rules.unique-variable-names/rules
             graphql-clj.validator.rules.unique-operation-names/rules
             graphql-clj.validator.rules.unique-fragment-names/rules
-            graphql-clj.validator.rules.unique-argument-names/rules]))
+            graphql-clj.validator.rules.unique-argument-names/statement-rules
+            graphql-clj.validator.rules.provided-non-null-arguments/rules]))
 
 (defn- validate [visit-fn]
   (try (visit-fn)
@@ -49,28 +55,27 @@
 (defn- validate-schema*
   "Do a 2 pass validation of a schema
    - First pass to add specs and validate that all types resolve.
-   - Second pass to apply all the validator rules.
-   There may be a clever way to avoid 2 passes...but for now it seems more interesting to be feature complete"
+   - Second pass to apply all the validator rules."
   [schema] ;; TODO inject introspection schema?
   (guard-parsed "schema" schema)
   (let [s (visitor/initial-state schema)
         {:keys [document state]} (visitor/visit-document schema s first-pass-rules)
-        second-pass (visitor/visit-document document state second-pass-rules)]
-    (assoc (:state second-pass) :schema (:document second-pass))))
+        second-pass (visitor/visit-document document state second-pass-rules-schema)]
+    (assoc-in second-pass [:state :schema] (:document second-pass))))
 
 (defn validate-statement*
   "Do a 2 pass validation of a statement"
-  [document' schema]
-  (guard-parsed "schema" schema)
+  [document' state]
+  (guard-parsed "schema" state)
   (guard-parsed "statement" document')
-  (let [s (assoc schema :statement-hash (hash document'))
+  (let [s (assoc state :statement-hash (hash document'))
         {:keys [document state]} (visitor/visit-document document' s first-pass-rules)]
-    (visitor/visit-document document state second-pass-rules)))
+    (visitor/visit-document document state second-pass-rules-statement)))
 
 ;; Public API
 
 (defn validate-schema [schema]
   (validate #(validate-schema* schema)))
 
-(defn validate-statement [document schema]
-  (validate #(validate-statement* document schema)))
+(defn validate-statement [document state]
+  (validate #(validate-statement* document state)))