+++ src/graphql_clj/visitor.clj
@@ -65,7 +65,7 @@
       (conj parent-path child-label)
       parent-path)))
 
-(def relevant-parent-keys #{:node-type :inner-type :type-name :spec :v/parent :v/path})
+(def relevant-parent-keys #{:node-type :inner-type :type-name :required :spec :v/parent :v/path})
 
 (defn- add-path [initial-state parentk parent child]
   (assoc child :v/parent (select-keys parent relevant-parent-keys) ;; TODO is this slow?