+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file. This change
 - Validate that fragments are on composite types
 - Make validation robust to order (can now reference types declared later in a statement or schema)
 - Validate unique variable, operation, input type, fragment, and argument names
+- Validate provided non-null arguments
 
 ## [0.1.17] - 2016-10-14
 - Implement preliminary version of FieldsOnCorrectType validation